+++ src/main/java/org/codinjutsu/tools/mongo/view/action/DropDatabaseAction.java
@@ -24,7 +24,11 @@ public class DropDatabaseAction extends AnAction implements DumbAware {
 
     @Override
     public void actionPerformed(AnActionEvent anActionEvent) {
-        mongoExplorerPanel.dropDatabase();
+        int result = JOptionPane.showConfirmDialog (null, String.format("Do you REALLY want to drop the '%s' database", mongoExplorerPanel.getSelectedDatabase().getName()),"Warning",JOptionPane.YES_NO_OPTION);
+
+        if(result == JOptionPane.YES_OPTION){
+            mongoExplorerPanel.dropDatabase();
+        }
     }
